InitializeDatabases/app/requirements.txt
install==1.3.5
mysql-connector-python==8.2.0
pip==23.3.2
protobuf==4.21.12
pyodbc==5.0.1
setuptools==58.1.0
